refactor: move board reads to board/api/board
